requirements.txt
tensorflow
tensorflow-addons
numpy
pandas
nltk
numpy
pandas
sentencepiece
sklearn
gensim
keras-tcn